+++ pkg/server/helpers.go
@@ -62,7 +62,7 @@ const (
 	relativeRootfsPath = "rootfs"
 	// defaultRuntime is the runtime to use in containerd. We may support
 	// other runtime in the future.
-	// defaultRuntime = "linux" // TODO defaulRuntime is currently unused
+	defaultRuntime = "linux"
 	// sandboxesDir contains all sandbox root. A sandbox root is the running
 	// directory of the sandbox, all files created for the sandbox will be
 	// placed under this directory.
