+++ pkg/server/image_remove.go
@@ -50,9 +50,8 @@ func (c *criContainerdService) RemoveImage(ctx context.Context, r *runtime.Remov
 		// return empty without error when image not found.
 		return &runtime.RemoveImageResponse{}, nil
 	}
-	// Also include repo digest, because if user pull image with digest,
-	// there will also be a corresponding repo digest reference.
-	for _, ref := range append(meta.RepoTags, meta.RepoDigests...) {
+	// Include all image references, including RepoTag, RepoDigest and id.
+	for _, ref := range append(append(meta.RepoTags, meta.RepoDigests...), meta.ID) {
 		// TODO(random-liu): Containerd should schedule a garbage collection immediately,
 		// and we may want to wait for the garbage collection to be over here.
 		err = c.imageStoreService.Delete(ctx, ref)
@@ -61,8 +60,7 @@ func (c *criContainerdService) RemoveImage(ctx context.Context, r *runtime.Remov
 		}
 		return nil, fmt.Errorf("failed to delete image reference %q for image %q: %v", ref, meta.ID, err)
 	}
-	err = c.imageMetadataStore.Delete(meta.ID)
-	if err != nil {
+	if err = c.imageMetadataStore.Delete(meta.ID); err != nil {
 		if metadata.IsNotExistError(err) {
 			return &runtime.RemoveImageResponse{}, nil
 		}
